ambassador-mqtt/requirements.txt
paho-mqtt==1.6.1
fastapi==0.104.1
uvicorn==0.24.0
